Implement citizens.accounts
This is the set of accounts for which the bank provides a comprehensive
list of transactions.

During aggregation all sets of accounts will be unioned so we will be
left with two sets: one of all known accounts, and its complement
unknown accounts. We will expect known-to-known transactions to arrive
in pairs since we have two sources for the same information. On the
other hand known-to-unknown transactions should only appear once, and
unknown-unknown transactions shouldn't exist.

One of the differences between known and unknown accounts is that we can
infer the change in balance of a known account since its inception by
summing incoming and negative outgoing transaction amounts. For unknown
accounts we can only infer the total change in their balance caused by
known accounts acting upon them.