Fix venmo floating error
An assertion caused venmo.transactions to fail on my system because
of a comparison between a decimal and float that forgot to take into
account floating point error from being unable to exactly represent
base 10 numbers. Fix the problem as close to the source as possible
by parsing decimals out of venmo's json instead of floats.